requirements.txt
flask_recaptcha
python-slugify